Source/StartUp button (#11) (#12)
* added http server and xl workbook

* removed genericfactory and cachedobject

* update workbook buttons

* update workbook buttons

* update workbook buttons

* update workbook

* update workbook

* Code Loading and Exporting
When opening the Workbook the code will be loaded, and when closing the
code will be written to files. Therefore the code has to lie next to the
worksheet.

* code is removed from xlsm on closing, for better versioning
of code and xlsm file

* update workbook incl. release v01

* update workbook incl. release v01

* removed workbook to make pull from pbrisk/master

* Merge branch 'master' of https://github.com/pbrisk/unicum into pbrisk-master

# Conflicts:
#	unicum/xl/workbook.xlsm

* update on workbook and server
better keys, values, items etc. for FactoryObject

* LineFeed/DoubleClick
Solved Problems due to parsing line feed chars on Windows
Doubleclick on Cells now checks if the conent is really an objectName of
an existing object, befor trying to write it on a new sheet

* SheetNames from Objects are checked
If the proposed name for a sheet from an Object does contain not allowed
characters or is too long (25+) the characters are first replaced and the
the name is shorted

* devtools update

* Merge branch 'jand' into master

* Merge branch 'jand' into master

* session handling

* session handling

* Added starting method for server,
some bugfixes and errorhandling

-Doubleclick on cells caused error when no session was running
-Statusbar is now properly set
-Clicking on the control"Source" allows the user to start a server, where
a PythonServerFile has to be supplied in the setup Sheet of the workbook

* minor changes on StartUp

* minor changes on StartUp

* better prepickle in server